Check cask and formulae versions
Enables upgrading casks with `brewp`. Modification to the jq filter that
will read also the casks field from the outdated packages json.

Fixed a bug, when there was only one current version, which was
represented by string value and not an array, previously jq filter was
expecting to always receive an array with at least one value.